clk: rockchip: rk3588: Support set_rate for SCLK_SDMMC_SAMPLE
Rockchip MMC driver is looking to set the clock based on sdmmc
node definition. For rk3588 the clock for setting the rate would
be SCLK_SDMMC_SAMPLE and which is indeed an unsupported clock.

This might not be true as the sdmmc node is added now for an
assumption to work however the actual node definition might
change once the Linux kernel supports it.

So, add SCLK_SDMMC_SAMPLE in set_rate and do-nothing for now
and which indeed enough to boot the SDMMC.

Signed-off-by: Jagan Teki <jagan@edgeble.ai>